Refactor N_COMPONENTS into common.py, and savefile functions out of it.
