+++ orbitx/common.py
@@ -4,19 +4,15 @@
 import atexit
 import logging
 import pytz
-import sys
 from io import StringIO
 from pathlib import Path
 from typing import NamedTuple, Optional
 
 import numpy
-import google.protobuf.json_format
 import vpython  # type: ignore
 
 from orbitx import orbitx_pb2 as protos
 from orbitx import strings
-from orbitx.data_structures.space import PhysicsState
-from orbitx.data_structures.eng_systems import N_COMPONENTS
 
 TIME_BETWEEN_NETWORK_UPDATES = 1.0
 
@@ -29,6 +25,17 @@ class TimeAcc(NamedTuple):
     desc: str
     # The acceleration above which this time acc starts to be too inaccurate.
     accurate_bound: float
+
+
+class OhmicVars(NamedTuple):
+    """
+    Encapsulates a Voltage, Current, and Resistance where V = IR.
+    This class is called 'Ohmic' because it's all the things for Ohm's law.
+    I don't think Ohmic is a word. I made it up. I like to think it sounds cool.
+    """
+    voltage: float
+    current: float
+    resistance: float
 
 
 # If you change the 'Pause' element of this list, change the corresponding
@@ -116,22 +123,18 @@ SRB_FULL = -1
 SRB_EMPTY = -2
 SRB_BURNTIME = 120  # 120s of burntime.
 
+
+# ---------------- Engineering-related constants ----------------------
+
+N_COMPONENTS = len(strings.COMPONENT_NAMES)
+N_COOLANT_LOOPS = len(strings.COOLANT_LOOP_NAMES)
+N_RADIATORS = len(strings.RADIATOR_NAMES)
+
 DANGEROUS_REACTOR_TEMP = 110  # We can change this later
 
 
-# ---------- Other runtime constants ----------
+# ---------- Other runtime functions and constants ----------
 PERF_FILE = 'flamegraph-data.log'
-
-if getattr(sys, 'frozen', False):
-    # We're running from a PyInstaller exe, use the path of the exe
-    PROGRAM_PATH = Path(sys.executable).parent
-elif sys.path[0] == '':
-    # We're running from a Python REPL. For information on what sys.path[0]
-    # means, read https://docs.python.org/3/library/sys.html#sys.path
-    # note path[0] == '' means Python is running as an interpreter.
-    PROGRAM_PATH = Path.cwd()
-else:
-    PROGRAM_PATH = Path(sys.path[0])
 
 
 def format_num(num: Optional[float], unit: str,
@@ -141,74 +144,6 @@ def format_num(num: Optional[float], unit: str,
     if num is None:
         return ''
     return '{:,.5g}'.format(round(num, ndigits=decimals)) + unit
-
-
-def savefile(name: str) -> Path:
-    return PROGRAM_PATH / 'data' / 'saves' / name
-
-
-def load_savefile(file: Path) -> 'PhysicsState':
-    """Loads the physics state represented by the input file.
-    If the input file is an OrbitX-style .json file, simply loads it.
-    If the input file is an OrbitV-style .rnd file, tries to interpret it
-    and also loads the adjacent STARSr file to produce a PhysicsState."""
-
-    # We shouldn't import orbitv_file_interface at the top of common.py, since
-    # common.py is imported by lots of modules and shouldn't circularly depend
-    # on anything.
-    from orbitx import orbitv_file_interface
-    physics_state: PhysicsState
-    logging.getLogger().info(f'Loading savefile {file.resolve()}')
-
-    assert isinstance(file, Path)
-    if file.suffix.lower() == '.rnd':
-        physics_state = \
-            orbitv_file_interface.clone_orbitv_state(file)
-
-    else:
-        if file.suffix.lower() != '.json':
-            logging.getLogger().warning(
-                f'{file} is not a .json file, trying to load it anyways.')
-
-        with open(file, 'r') as f:
-            data = f.read()
-        read_state = protos.PhysicalState()
-        google.protobuf.json_format.Parse(data, read_state)
-
-        if len(read_state.engineering.components) == 0:
-            # We allow savefiles to not specify any components.
-            # If we see this, create a list of empty components in the protobuf before parsing it.
-            empty_components = [protos.EngineeringState.Component()] * N_COMPONENTS
-            read_state.engineering.components.extend(empty_components)
-
-        physics_state = PhysicsState(None, read_state)
-
-    if physics_state.timestamp == 0:
-        physics_state.timestamp = DEFAULT_INITIAL_TIMESTAMP
-    if physics_state.time_acc == 0:
-        physics_state.time_acc = DEFAULT_TIME_ACC.value
-    if physics_state.reference == '':
-        physics_state.reference = DEFAULT_REFERENCE
-    if physics_state.target == '':
-        physics_state.target = DEFAULT_TARGET
-    if physics_state.srb_time == 0:
-        physics_state.srb_time = SRB_FULL
-
-    return physics_state
-
-
-def write_savefile(state: 'PhysicsState', file: Path):
-    """Writes state to the specified savefile path (use common.savefile to get
-    a savefile path in data/saves/). Returns a possibly-different path that it
-    was saved under."""
-    if file.suffix.lower() != '.json':
-        # Ensure a .json suffix.
-        file = file.parent / (file.name + '.json')
-    logging.getLogger().info(f'Saving to savefile {file.resolve()}')
-    with open(file, 'w') as outfile:
-        outfile.write(
-            google.protobuf.json_format.MessageToJson(state.as_proto()))
-    return file
 
 
 def start_flamegraphing():
