+++ orbitx/graphics/eng/tkinter_widgets.py
@@ -10,7 +10,8 @@ import tkinter as tk
 from PIL import Image, ImageTk
 from typing import List, Optional, Callable
 
-from orbitx.data_structures import EngineeringState, Request
+from orbitx.network import Request
+from orbitx.data_structures.engineering import EngineeringState
 from orbitx.programs import hab_eng
 from orbitx import strings
