+++ orbitx/programs/flight_training.py
@@ -8,6 +8,7 @@ from orbitx import common
 from orbitx import physics
 from orbitx import programs
 from orbitx.graphics.flight import flight_gui
+from orbitx.data_structures import savefile
 
 log = logging.getLogger('orbitx')
 
@@ -24,7 +25,7 @@ argument_parser = argparse.ArgumentParser(
 argument_parser.add_argument(
     'loadfile', type=str, nargs='?', default='OCESS.json',
     help=(
-        f'Name of the savefile to load, relative to {common.savefile(".")}. '
+        f'Name of the savefile to load, relative to {savefile.full_path(".")}. '
         'Should be a .json savefile written by OrbitX. '
         'Can also read OrbitV .RND savefiles.')
 )
@@ -36,9 +37,9 @@ def main(args: argparse.Namespace):
         loadfile = Path(args.loadfile)
     else:
         # Take paths relative to 'data/saves/'
-        loadfile = common.savefile(args.loadfile)
+        loadfile = savefile.full_path(args.loadfile)
 
-    physics_engine = physics.PhysicsEngine(common.load_savefile(loadfile))
+    physics_engine = physics.PhysicsEngine(savefile.load_savefile(loadfile))
     initial_state = physics_engine.get_state()
 
     gui = flight_gui.FlightGui(
