+++ orbitx/programs/mc_flight.py
@@ -3,10 +3,9 @@ import atexit
 import logging
 
 from orbitx import common
-from orbitx import network
 from orbitx import programs
 from orbitx.graphics.flight import flight_gui
-from orbitx.data_structures import Request
+from orbitx.network import Request, NetworkedStateClient
 
 log = logging.getLogger('orbitx')
 
@@ -31,7 +30,7 @@ def main(args: argparse.Namespace):
     networking = True  # Whether data is requested over the network
 
     log.info(f'Connecting to physics server {args.physics_server}.')
-    lead_server_connection = network.NetworkedStateClient(
+    lead_server_connection = NetworkedStateClient(
         Request.MC_FLIGHT, args.physics_server)
     state = lead_server_connection.get_state()
     physics_engine = lead_server_connection._caching_physics_engine
